student-campaign-system/requirements.txt
Flask>=2.3.0,<3.0.0
Flask-CORS>=4.0.0
SQLAlchemy>=2.0.0,<3.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.12.0
flake8>=6.0.0
